backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1

# WebSocket support
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Redis (for GEO and caching)
redis==5.0.0

# Celery (async task queue)
celery==5.3.4
python-dotenv==1.0.0
Pillow==10.4.0
